Fix minor issues
